Split ToggleSection into Primary and Secondary appearances
