+++ src/client/components/ToggleSection/usage.md
@@ -3,6 +3,4 @@ Toggle section
 
 ### Description
 
-This is to be used for toggling groups of filters, this does not replace the "details" component in the [GDS design system](https://design-system.service.gov.uk/components/details/)
-
-
+This is to be used for toggling groups, this does not replace the "details" component in the [GDS design system](https://design-system.service.gov.uk/components/details/).
